+++ libs/components/src/lib/components/forms/components/input/stories/components/textarea-input-storybook.component.ts
@@ -1,0 +1,26 @@
+import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core';
+import { FormControl, ReactiveFormsModule } from '@angular/forms';
+import { InputImports, LabelComponent } from '../../..';
+
+@Component({
+  selector: 'et-sb-textarea-input',
+  template: `
+    <et-input-field [formControl]="fg">
+      <et-textarea-input [cols]="cols" [rows]="rows" />
+      <et-label>Textarea input</et-label>
+    </et-input-field>
+  `,
+  standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None,
+  imports: [InputImports, LabelComponent, ReactiveFormsModule],
+})
+export class StorybookTextareaInputComponent {
+  fg = new FormControl();
+
+  @Input()
+  cols: number | null = null;
+
+  @Input()
+  rows: number | null = null;
+}